% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcbs.R
\name{dcbs.branch}
\alias{dcbs.branch}
\title{Growing a branch for DCBS algorithm}
\usage{
dcbs.branch(input, phi, s, e, trim)
}
\arguments{
\item{input}{input data matrix, with each row representing the component time series or their transformation}

\item{phi, trim}{see \code{\link{dcbs.alg}}}

\item{s, e}{start and end of the interval of consideration at a given iteration}
}
\value{
a vector containing the information about the branch, including the location of the estimated change-point and the corresponding test statistic
}
\description{
Grow a branch of the binary tree for the Double CUSUM Binary Segmentation without thresholding
}
\keyword{internal}
