% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbs.R
\name{search.b}
\alias{search.b}
\title{Searching for a change-point in a branch}
\usage{
search.b(stat, trim)
}
\arguments{
\item{stat}{aggregated CUSUM statistics}

\item{trim}{see \code{\link{sbs.alg}}}
}
\value{
a list containing
   \item{b}{a possible location of the change-point}
   \item{test.stat}{test statistic}
   \item{FLAG}{should the branch be grown?}
}
\description{
Searches for a change-point in a branch of a binary tree grown by the Sparsified Binary Segmentation
}
\keyword{internal}
