% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{tri.kern}
\alias{tri.kern}
\title{Flat-top kernel}
\usage{
tri.kern(h, len = 0)
}
\arguments{
\item{h}{bandwidth}

\item{len}{desired length of the vector containing the flat-top kernel}
}
\value{
a vector containing half of the flat-top kernel
}
\description{
Flat-top kernel
}
\keyword{internal}
